toolkit/content/tests/chrome/bug331215_window.xul

<?xml version="1.0"?>

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->

<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet
  href="chrome://mochikit/content/tests/SimpleTest/test.css"
  type="text/css"?>

<window id="331215test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        width="600"
        height="600"
        onload="SimpleTest.executeSoon(startTest);"
        title="331215 test">

  <script type="application/javascript"><![CDATA[
    const {interfaces: Ci, classes: Cc, results: Cr, utils: Cu} = Components;
    Cu.import("resource://gre/modules/Task.jsm");
    Cu.import("resource://testing-common/ContentTask.jsm");
    ContentTask.setTestScope(window.opener.wrappedJSObject);

    var gFindBar = null;
    var gBrowser;

    var imports = ["SimpleTest", "ok", "info"];
    for (var name of imports) {
      window[name] = window.opener.wrappedJSObject[name];
    }

    function startTest() {
      Task.spawn(function* () {
        gFindBar = document.getElementById("FindToolbar");
        for (let browserId of ["content", "content-remote"]) {
          yield startTestWithBrowser(browserId);
        }
      }).then(() => {
        window.close();
        SimpleTest.finish();
      });
    }

    function* startTestWithBrowser(browserId) {
      info("Starting test with browser '" + browserId + "'");
      gBrowser = document.getElementById(browserId);
      gFindBar.browser = gBrowser;
      let promise = ContentTask.spawn(gBrowser, null, function* () {
        return new Promise(resolve => {
          addEventListener("DOMContentLoaded", function listener() {
            removeEventListener("DOMContentLoaded", listener);
            resolve();
          });
        });
      });
      gBrowser.loadURI("data:text/plain,latest");
      yield promise;
      yield onDocumentLoaded();
    }

    function* onDocumentLoaded() {
      document.getElementById("cmd_find").doCommand();
      yield enterStringIntoFindField("test");
      document.commandDispatcher
              .getControllerForCommand("cmd_moveTop")
              .doCommand("cmd_moveTop");
      yield enterStringIntoFindField("l");
      ok(gFindBar._findField.getAttribute("status") == "notfound",
         "Findfield status attribute should have been 'notfound'" +
         " after entering ltest");              
      yield enterStringIntoFindField("a");
      ok(gFindBar._findField.getAttribute("status") != "notfound",
         "Findfield status attribute should not have been 'notfound'" +
         " after entering latest");
    }

    function* enterStringIntoFindField(aString) {
      for (let i = 0; i < aString.length; i++) {      
        let event = document.createEvent("KeyEvents");
        let promise = new Promise(resolve => {
          let listener = {
            onFindResult: function() {
              gFindBar.browser.finder.removeResultListener(listener);
              resolve();
            }
          };
          gFindBar.browser.finder.addResultListener(listener);
        });
        event.initKeyEvent("keypress", true, true, null, false, false,
                           false, false, 0, aString.charCodeAt(i));
        gFindBar._findField.inputField.dispatchEvent(event);
        yield promise;
      }
    }
  ]]></script>

  <commandset>
    <command id="cmd_find" oncommand="document.getElementById('FindToolbar').onFindCommand();"/>
  </commandset>
  <browser type="content-primary" flex="1" id="content" src="about:blank"/>
  <browser type="content-primary" flex="1" id="content-remote" remote="true" src="about:blank"/>
  <findbar id="FindToolbar" browserid="content"/>
</window>
